ENH: limit memory usage
Slight rework of mask_and_reduce to limit memory usage.

The trick is to delay as much as possible loading of data. It is critical
partly because things don't always get garbage collected as well as they
should, and data sticks around (nifti images may have references on data
after it's loaded, for instance).

The solution here is not satisfactory in the long run, but at least
it is now in constant memory usage, as before the big dict learning
refactoring.